--- packages/form-render/src/widgets/antd/list.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function map({ children }) {
-  return <div className="w-100">{children}</div>;
-}
