+++ packages/form-render/src/widgets/antd/list/CardList.less
@@ -1,7 +1,4 @@
 .fr-container {
-  .fr-card-list {
-  }
-
   .fr-card-item {
     border: 1px solid rgba(0, 0, 0, 0.1);
     border-radius: 4px;
