+++ packages/form-render/src/widgets/antd/list/TableList.js
@@ -1,17 +1,16 @@
 /* eslint-disable jsx-a11y/anchor-is-valid */
 import React from 'react';
-import Core from '../../index';
-import { Button, Popconfirm, Table } from 'antd';
-import { useVT } from 'virtualizedtableforantd4';
+import { Button, Table, Popconfirm } from 'antd';
 import { ArrowUpOutlined, ArrowDownOutlined } from '@ant-design/icons';
 
 const FIELD_LENGTH = 170;
 
-const VirtualList = ({
+const TableList = ({
   displayList = [],
   dataIndex,
   children,
   deleteItem,
+  copyItem,
   addItem,
   moveItemUp,
   moveItemDown,
@@ -19,11 +18,17 @@ const VirtualList = ({
   schema,
   listData,
   changeList,
+  Field,
 }) => {
   const { props = {}, itemProps = {} } = schema;
-  const { scrollY = 600, ...rest } = props;
+  const { buttons, ...columnProps } = itemProps;
+  const { pagination = {}, ...rest } = props;
 
-  const [vt, set_components] = useVT(() => ({ scroll: { y: scrollY } }), []);
+  const paginationConfig = pagination && {
+    size: 'small',
+    hideOnSinglePage: true,
+    ...pagination,
+  };
 
   const dataSource = displayList.map((item, idx) => {
     return { index: idx };
@@ -34,7 +39,6 @@ const VirtualList = ({
     const schema = (item && item.schema) || {};
     return {
       dataIndex: child,
-      width: FIELD_LENGTH,
       title: schema.required ? (
         <>
           <span className="fr-label-required"> *</span>
@@ -43,11 +47,12 @@ const VirtualList = ({
       ) : (
         schema.title
       ),
+      width: FIELD_LENGTH,
       render: (value, record, index) => {
         // Check: record.index 似乎是antd自己会给的，不错哦
         const childIndex = [...dataIndex, record.index];
         return (
-          <Core
+          <Field
             hideTitle={true}
             displayType="inline"
             key={index.toString()}
@@ -56,18 +61,28 @@ const VirtualList = ({
           />
         );
       },
+      ...columnProps,
     };
   });
 
-  if (!props.hideDelete || Array.isArray(itemProps.buttons)) {
+  if (
+    !props.hideDelete ||
+    !props.hideAdd ||
+    !props.hideCopy ||
+    !props.hideMove
+  ) {
     columns.push({
       title: '操作',
       key: '$action',
       fixed: 'right',
       width: 120,
-      render: (value, record, idx) => {
+      render: (value, record) => {
+        const idx = record.index;
         return (
-          <>
+          <div>
+            {!props.hideAdd && !props.hideCopy && (
+              <a onClick={() => copyItem(idx)}>复制</a>
+            )}
             {!props.hideDelete && (
               <Popconfirm
                 title="确定删除?"
@@ -75,7 +90,7 @@ const VirtualList = ({
                 okText="确定"
                 cancelText="取消"
               >
-                <a>删除</a>
+                <a style={{ marginLeft: 8 }}>删除</a>
               </Popconfirm>
             )}
             {!props.hideMove && (
@@ -90,34 +105,7 @@ const VirtualList = ({
                 />
               </>
             )}
-            {Array.isArray(itemProps.buttons)
-              ? itemProps.buttons.map((item, idx) => {
-                  const { callback, text, html } = item;
-                  let onClick = () => {};
-                  if (typeof window[callback] === 'function') {
-                    onClick = () => {
-                      window[callback]({
-                        value: listData,
-                        onChange: changeList,
-                        schema,
-                      });
-                    };
-                  }
-                  return (
-                    <a
-                      key={idx.toString()}
-                      style={{ marginLeft: 8 }}
-                      size="small"
-                      onClick={onClick}
-                    >
-                      <span
-                        dangerouslySetInnerHTML={{ __html: html || text }}
-                      />
-                    </a>
-                  );
-                })
-              : null}
-          </>
+          </div>
         );
       },
     });
@@ -134,8 +122,13 @@ const VirtualList = ({
         {Array.isArray(props.buttons)
           ? props.buttons.map((item, idx) => {
               const { callback, text, html } = item;
-
-              let onClick = () => {};
+              let onClick = () => {
+                console.log({
+                  value: listData,
+                  onChange: changeList,
+                  schema,
+                });
+              };
               if (typeof window[callback] === 'function') {
                 onClick = () => {
                   window[callback]({
@@ -158,19 +151,17 @@ const VirtualList = ({
             })
           : null}
       </div>
-
       <Table
-        rowKey="index"
-        scroll={{ y: scrollY, x: 'max-content' }}
-        components={vt}
-        size="small"
+        scroll={{ x: 'max-content' }}
         columns={columns}
         dataSource={dataSource}
-        pagination={false}
+        rowKey="index"
+        size="small"
+        pagination={paginationConfig}
         {...rest}
       />
     </>
   );
 };
 
-export default VirtualList;
+export default TableList;
