+++ packages/form-render/src/widgets/antd/list/VirtualList.js
@@ -1,17 +1,16 @@
 /* eslint-disable jsx-a11y/anchor-is-valid */
 import React from 'react';
-import Core from '../../index';
-import { Button, Table, Popconfirm } from 'antd';
+import { Button, Popconfirm, Table } from 'antd';
+import { useVT } from 'virtualizedtableforantd4';
 import { ArrowUpOutlined, ArrowDownOutlined } from '@ant-design/icons';
 
 const FIELD_LENGTH = 170;
 
-const TableList = ({
+const VirtualList = ({
   displayList = [],
   dataIndex,
   children,
   deleteItem,
-  copyItem,
   addItem,
   moveItemUp,
   moveItemDown,
@@ -19,16 +18,12 @@ const TableList = ({
   schema,
   listData,
   changeList,
+  Field,
 }) => {
   const { props = {}, itemProps = {} } = schema;
-  const { buttons, ...columnProps } = itemProps;
-  const { pagination = {}, ...rest } = props;
+  const { scrollY = 600, ...rest } = props;
 
-  const paginationConfig = pagination && {
-    size: 'small',
-    hideOnSinglePage: true,
-    ...pagination,
-  };
+  const [vt, set_components] = useVT(() => ({ scroll: { y: scrollY } }), []);
 
   const dataSource = displayList.map((item, idx) => {
     return { index: idx };
@@ -39,6 +34,7 @@ const TableList = ({
     const schema = (item && item.schema) || {};
     return {
       dataIndex: child,
+      width: FIELD_LENGTH,
       title: schema.required ? (
         <>
           <span className="fr-label-required"> *</span>
@@ -47,12 +43,11 @@ const TableList = ({
       ) : (
         schema.title
       ),
-      width: FIELD_LENGTH,
       render: (value, record, index) => {
         // Check: record.index 似乎是antd自己会给的，不错哦
         const childIndex = [...dataIndex, record.index];
         return (
-          <Core
+          <Field
             hideTitle={true}
             displayType="inline"
             key={index.toString()}
@@ -61,28 +56,18 @@ const TableList = ({
           />
         );
       },
-      ...columnProps,
     };
   });
 
-  if (
-    !props.hideDelete ||
-    !props.hideAdd ||
-    !props.hideCopy ||
-    !props.hideMove
-  ) {
+  if (!props.hideDelete || Array.isArray(itemProps.buttons)) {
     columns.push({
       title: '操作',
       key: '$action',
       fixed: 'right',
       width: 120,
-      render: (value, record) => {
-        const idx = record.index;
+      render: (value, record, idx) => {
         return (
-          <div>
-            {!props.hideAdd && !props.hideCopy && (
-              <a onClick={() => copyItem(idx)}>复制</a>
-            )}
+          <>
             {!props.hideDelete && (
               <Popconfirm
                 title="确定删除?"
@@ -90,7 +75,7 @@ const TableList = ({
                 okText="确定"
                 cancelText="取消"
               >
-                <a style={{ marginLeft: 8 }}>删除</a>
+                <a>删除</a>
               </Popconfirm>
             )}
             {!props.hideMove && (
@@ -105,7 +90,34 @@ const TableList = ({
                 />
               </>
             )}
-          </div>
+            {Array.isArray(itemProps.buttons)
+              ? itemProps.buttons.map((item, idx) => {
+                  const { callback, text, html } = item;
+                  let onClick = () => {};
+                  if (typeof window[callback] === 'function') {
+                    onClick = () => {
+                      window[callback]({
+                        value: listData,
+                        onChange: changeList,
+                        schema,
+                      });
+                    };
+                  }
+                  return (
+                    <a
+                      key={idx.toString()}
+                      style={{ marginLeft: 8 }}
+                      size="small"
+                      onClick={onClick}
+                    >
+                      <span
+                        dangerouslySetInnerHTML={{ __html: html || text }}
+                      />
+                    </a>
+                  );
+                })
+              : null}
+          </>
         );
       },
     });
@@ -122,13 +134,8 @@ const TableList = ({
         {Array.isArray(props.buttons)
           ? props.buttons.map((item, idx) => {
               const { callback, text, html } = item;
-              let onClick = () => {
-                console.log({
-                  value: listData,
-                  onChange: changeList,
-                  schema,
-                });
-              };
+
+              let onClick = () => {};
               if (typeof window[callback] === 'function') {
                 onClick = () => {
                   window[callback]({
@@ -151,17 +158,19 @@ const TableList = ({
             })
           : null}
       </div>
+
       <Table
-        scroll={{ x: 'max-content' }}
+        rowKey="index"
+        scroll={{ y: scrollY, x: 'max-content' }}
+        components={vt}
+        size="small"
         columns={columns}
         dataSource={dataSource}
-        rowKey="index"
-        size="small"
-        pagination={paginationConfig}
+        pagination={false}
         {...rest}
       />
     </>
   );
 };
 
-export default TableList;
+export default VirtualList;
